requirements.txt
requests>=2.28.0
html2text>=2020.1.16
PyYAML>=6.0
anthropic>=0.18.0
